refactor: use Typescript
This is a relatively large commit to update the project to use Typescript.

The core of this project is to transform objects from one type to the other: from the V8 format to the Istambul format. Typescript would allow to formally document the input and output types, thus helping new maintainers to be more confident when doing changes. Beyond this, it brings the benefits of type definitions to consumers of the library (completion, type checking, etc).

This first commit does not do any change in the behavior of the code, it does not even type anything. It just adds Typescript and converts the files.

It adds `tslint` as a linter. This commit still results in lint errors but they will be fixed in a subsequent commit.

It still uses TAP but code coverage no longer works. A subsequent commit should fix it by running `tap` on the build directory, or switching to `mocha`.

Even at this minimal stage, it caught an error in `CovBranch` were the undefined property `line` was used.